# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Puesto de Salud Bohórquez",
    "description": "API para gestionar pacientes, doctores, servicios, consultorios y citas.",
    "version": "1.0.0"
  },
  "servers": [
    { "url": "http://localhost:5000" }
  ],
...
  "paths": {
    "/pacientes": {
      "get": {
        "tags": ["Pacientes"],
        "summary": "Listar todos los pacientes",
        "responses": {
          "200": {
            "description": "Lista de pacientes",
            "content": {
              "application/json": {
                "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Paciente" } }
              }
            }
          }
        }
      },

      "post": {
        "tags": ["Pacientes"],
        "summary": "Registrar un nuevo paciente",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/PacienteCreate" }
            }
          }
        },
        "responses": {
          "201": { "description": "Paciente registrado" },
          "400": { "description": "Datos incompletos o inválidos" }
        }
      }
    },
    "/pacientes/{id}": {
      "put": {
        "tags": ["Pacientes"],
        "summary": "Actualizar un paciente",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/PacienteCreate" }
            }
          }
        },
        "responses": {
          "200": { "description": "Paciente actualizado" },
          "404": { "description": "Paciente no encontrado" }
        }
      },

      "delete": {
        "tags": ["Pacientes"],
        "summary": "Eliminar un paciente",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": { "description": "Paciente eliminado" },
          "404": { "description": "Paciente no encontrado" }
        }
      }
    },
    "/doctores": {
      "get": {
        "tags": ["Doctores"],
        "summary": "Listar doctores",
        "responses": {
          "200": {
            "description": "Lista de doctores",
            "content": {
              "application/json": {
                "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Doctor" } }
              }
            }
          }
        }
      },

      "post": {
        "tags": ["Doctores"],
        "summary": "Registrar un doctor",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/DoctorCreate" }
            }
          }
        },
        "responses": {
          "201": { "description": "Doctor registrado" },
          "400": { "description": "Datos inválidos" }
        }
      }
    },
    "/doctores/{id}": {
      "put": {
        "tags": ["Doctores"],
        "summary": "Actualizar un doctor",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/DoctorCreate" }
            }
          }
        },
        "responses": {
          "200": { "description": "Doctor actualizado" },
          "404": { "description": "Doctor no encontrado" }
        }
      },

      "delete": {
        "tags": ["Doctores"],
        "summary": "Eliminar un doctor",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": { "description": "Doctor eliminado" },
          "404": { "description": "Doctor no encontrado" }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Paciente": {
        "type": "object",
        "properties": {
          "id_paciente": { "type": "integer" },
          "documento": { "type": "string" },
          "nombre": { "type": "string" },
          "fecha_nacimiento": { "type": "string" },
          "telefono": { "type": "string" },
          "direccion": { "type": "string" },
          "correo": { "type": "string" }
        }
      },
      "PacienteCreate": {
        "type": "object",
        "required": ["documento", "nombre", "fecha_nacimiento", "telefono", "direccion", "correo"],
        "properties": {
          "documento": { "type": "string" },
          "nombre": { "type": "string" },
          "fecha_nacimiento": { "type": "string" },
          "telefono": { "type": "string" },
          "direccion": { "type": "string" },
          "correo": { "type": "string" }
        }
      },
      "Doctor": {
        "type": "object",
        "properties": {
          "id_profesional": { "type": "integer" },
          "nombre": { "type": "string" },
          "especialidad": { "type": "string" },
          "telefono": { "type": "string" },
          "correo": { "type": "string" }
        }
      },
      "DoctorCreate": {
        "type": "object",
        "required": ["nombre", "especialidad", "telefono", "correo"],
        "properties": {
          "nombre": { "type": "string" },
          "especialidad": { "type": "string" },
          "telefono": { "type": "string" },
          "correo": { "type": "string" }
        }
      },
...
    }
  }
}